move cascade tests to their own file
